+++ beets/library.py
@@ -217,6 +217,7 @@ class Item(object):
         """Writes the item's metadata to the associated file.
         """
         f = MediaFile(syspath(self.path))
+        plugins.send('write', item=self, mf=f)
         for key in ITEM_KEYS_WRITABLE:
             setattr(f, key, getattr(self, key))
         f.save()